+++ components/decks/DeckDialog.tsx
@@ -38,6 +38,7 @@ import { WordList } from "@/components/decks/WordList";
 import type { WordRowValue } from "@/components/decks/WordRow";
 import { AiImportPanel } from "@/components/decks/AiImportPanel";
 import type { ParsedAiDeck } from "@/lib/parseAiDeckResponse";
+import { mergeDeckImport } from "@/lib/deckImport";
 
 interface DeckDialogProps {
   deck?: Deck; // deck이 있으면 수정 모드, 없으면 생성 모드
@@ -256,97 +257,9 @@ export function DeckDialog({ deck, children }: DeckDialogProps) {
         setDescription(parsed.description);
       }
 
-      setFormState((prev) => {
-        // categories: 합집합 (기존 순서 유지 + 새 항목 append)
-        const existingCats = new Set(prev.categories);
-        const newCats = parsed.categories.filter((c) => !existingCats.has(c));
-        const mergedCategories = [...prev.categories, ...newCats];
-
-        // 끝의 빈 행 분리 (모두 제거 후 마지막에 1개만 다시 추가)
-        const baseWords = [...prev.words];
-        while (
-          baseWords.length > 0 &&
-          baseWords[baseWords.length - 1].word.trim().length === 0 &&
-          baseWords[baseWords.length - 1].tags.length === 0
-        ) {
-          baseWords.pop();
-        }
-
-        // 정규화된 word를 키로 사용 (incoming.word와 비교 일관성)
-        const byWord = new Map<string, { idx: number; row: WordRowValue }>();
-        for (let i = 0; i < baseWords.length; i++) {
-          const row = baseWords[i];
-          const key = row.word.trim().toLowerCase();
-          if (key) byWord.set(key, { idx: i, row });
-        }
-
-        const merged: WordRowValue[] = [...baseWords];
-        for (const incoming of parsed.words) {
-          const existing = byWord.get(incoming.word);
-          if (existing) {
-            const updated: WordRowValue = {
-              ...existing.row,
-              tags: Array.from(
-                new Set([...existing.row.tags, ...incoming.tags])
-              ),
-            };
-            merged[existing.idx] = updated;
-            byWord.set(incoming.word, { idx: existing.idx, row: updated });
-          } else {
-            const row: WordRowValue = {
-              id: nanoid(),
-              word: incoming.word,
-              tags: incoming.tags,
-            };
-            const idx = merged.length;
-            merged.push(row);
-            byWord.set(incoming.word, { idx, row });
-          }
-        }
-
-        // 끝에 빈 행 1개 유지 (사용자가 추가 입력하기 쉽게)
-        merged.push(makeRow());
-
-        const incomingHasTags = parsed.words.some((w) => w.tags.length > 0);
-        const willTurnOn =
-          !prev.usesCategories &&
-          (mergedCategories.length > 0 || incomingHasTags);
-
-        if (willTurnOn) {
-          // OFF → ON 자동 전환: 기존 stash 복원 (handleToggleCategories(true) 미러링)
-          const restoredWords = merged.map((row) => {
-            const stashed = prev.hiddenWordTags[row.id];
-            if (stashed && stashed.length > 0 && row.tags.length === 0) {
-              return { ...row, tags: stashed };
-            }
-            return row;
-          });
-          // 기존 hiddenCategories 중 mergedCategories에 없는 것은 복원
-          const mergedSet = new Set(mergedCategories);
-          const restoredCategories = [...mergedCategories];
-          for (const c of prev.hiddenCategories) {
-            if (!mergedSet.has(c)) {
-              mergedSet.add(c);
-              restoredCategories.push(c);
-            }
-          }
-          return {
-            ...prev,
-            categories: restoredCategories,
-            words: restoredWords,
-            usesCategories: true,
-            hiddenCategories: [],
-            hiddenWordTags: {},
-          };
-        }
-
-        return {
-          ...prev,
-          categories: mergedCategories,
-          words: merged,
-          // 이미 ON이면 그대로 유지, 아니면 OFF 유지 (stash도 그대로)
-        };
-      });
+      setFormState((prev) =>
+        mergeDeckImport(prev, parsed, script) as DeckFormState
+      );
 
       const droppedCount = parsed.droppedWords.length;
       toast.success(`${parsed.words.length}개 단어를 추가했습니다.`);
@@ -356,7 +269,7 @@ export function DeckDialog({ deck, children }: DeckDialogProps) {
         );
       }
     },
-    [name, description]
+    [name, description, script]
   );
 
   const handleToggleCategories = useCallback((checked: boolean) => {
